pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "neutts-fastapi"
version = "0.1.0"
description = "OpenAI-compatible FastAPI server for NeuTTS text-to-speech"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = "MIT"
dependencies = [
    "neutts>=1.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.0.0",
    "av>=12.0.0",
    "loguru>=0.7.0",
    "psutil>=5.9.0",
    "python-multipart>=0.0.7",
    "soundfile>=0.13.0",
    "websockets>=12.0",
    "setuptools>=70.0.0",  # Required by perth (resemble-perth) for pkg_resources
]

[project.optional-dependencies]
gpu = ["neutts[all]", "onnxruntime-gpu", "torch>=2.6.0"]
cpu = ["neutts[all]", "onnxruntime"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
